requirements.txt
django-bootstrap5==25.1
onnxruntime==1.21.0
pip-chill==1.0.3
python-decouple==3.8
rembg==2.0.64
svglib==1.5.1
